Remove artifact_signature for the helm chart for now
We still store some old artifact versions in the db and in github that are not
signed - by default minder stores up to 30 days worth of artifacts. At the
same time, we don't use any floating tag like latest so it's problematic
to pin a single tag.

Since the failure wouldn't tell us much, let's remove the rule for now.