Remove the recently added default
Our ruletype update validation throws an error if you modify the default
field in any way which is probably incorrect as the default is really
not used except as advisory, but let's not break ruletype apply.

I added the default recently, but luckily the update didn't make it to
the healtcheck yet.