Exclude the action name only not the full uses in actions_check_pinned_tags
Our example profiles exclude just the action name because that's what
frizbee uses internally to exclude:
https://github.com/stacklok/minder-rules-and-profiles/blob/main/profiles/github/stacklok-health-check.yaml#L29

But the rego check in the actions_check_pinned_tags rule was excluding
on the full value of uses. This meant that the remediation was always
called even if it didn't produce anything and the exclude didn't work.