Update rule configuration (#179)
* Remove branch name requirement in rules

The branch protection rules have a sensible default of the remote HEAD
branch. As a result, they no longer _require_ the branch name. Right
now, callers must specify the branch name as an empty string (since it's
required), but we should instead just remove the requirement entirely.

* Remove filename requirement in rules

The security insights rules have a sensible default of
`SECURITY-INSIGHTS.yml`, thus, we should not require the filename.

* Remove schedule_interval from dependabot rule

The dependabot rule does not use the schedule_interval option; it should
not be required any longer.

* Remove filename requirement in security rules

The security policy rule has a sensible default of `SECURITY.md`,
thus, we should not require the filename.