Add OSI and GitHub data sources. (#219)
This change adds two data sources useful to implement OSPS Baseline
checks.

OSI data source is pretty simple and static, and can be used to verify
if a given license is allowed by either OSI or FSF.

GitHub data source is useful to get the license of a given public
repository as detected by GitHub. Note that this currently only works
for public repositories.